discordbot/requirements.txt
aiohttp==3.8.4
async-timeout
asyncio-extras==1.3.2
discord.py==2.2.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-socketio[asyncio_client]==5.8.0
quart==0.18.3
requests==2.28.2
redis[hiredis]==4.5.4